+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 - Add support for 1.21.4
 - Add support for Java 23
+- Fix "logged in from another location" issue on Paper (see GH #296)